scc-top/requirements.txt
pyyaml>=6.0.1
pytest>=8.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9

